+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './typed-fetch';
+export * from './typed-fetch/api-operations';